+++ ruoyi-admin/src/main/java/com/ruoyi/web/controller/platform/EconTypeController.java
@@ -1,26 +1,29 @@
 package com.ruoyi.web.controller.platform;
 
-import com.ruoyi.common.annotation.Log;
-import com.ruoyi.common.core.controller.BaseController;
-import com.ruoyi.common.core.domain.AjaxResult;
-import com.ruoyi.common.core.page.TableDataInfo;
-import com.ruoyi.common.enums.BusinessType;
-import com.ruoyi.common.utils.poi.ExcelUtil;
-import com.ruoyi.province.platform.domain.EconType;
-import com.ruoyi.province.platform.service.IEconTypeService;
+import java.util.List;
 import org.apache.shiro.authz.annotation.RequiresPermissions;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Controller;
 import org.springframework.ui.ModelMap;
-import org.springframework.web.bind.annotation.*;
-
-import java.util.List;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.PostMapping;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.ResponseBody;
+import com.ruoyi.common.annotation.Log;
+import com.ruoyi.common.enums.BusinessType;
+import com.ruoyi.province.platform.domain.EconType;
+import com.ruoyi.province.platform.service.IEconTypeService;
+import com.ruoyi.common.core.controller.BaseController;
+import com.ruoyi.common.core.domain.AjaxResult;
+import com.ruoyi.common.utils.poi.ExcelUtil;
+import com.ruoyi.common.core.page.TableDataInfo;
 
 /**
  * 经济类型Controller
  * 
  * @author dalin
- * @date 2020-12-08
+ * @date 2020-12-09
  */
 @Controller
 @RequestMapping("/platform/econtype")
@@ -90,7 +93,7 @@ public class EconTypeController extends BaseController
      * 修改经济类型
      */
     @GetMapping("/edit/{econId}")
-    public String edit(@PathVariable("econId") Integer econId, ModelMap mmap)
+    public String edit(@PathVariable("econId") Long econId, ModelMap mmap)
     {
         EconType econType = econTypeService.selectEconTypeById(econId);
         mmap.put("econType", econType);
